coccinelle: flag HZ dependent constants being passed for jiffies
A large set of calls in the kernel takes jiffies as an argument - passing
in a constant makes the timeout HZ dependent which is wrong in all cases
found by this spatch and probably is wrong for any call that expects
jiffies with the exception of passing in 1 (minimum delay).

Checking for constants only yields many false positives so they are
filtered for digits only. A numeric value of 1 is though commonly in use
for "shortest possible delay" so those cases are treated as false
positives as well and not reported.

Change-Id: I1f3184352c9ef572dd7113988afa3bfc2e28b041
Signed-off-by: Nicholas Mc Guire <hofrat@osadl.org>