Make msm_serial_hs RT to improve bluetooth performance
msm_serial_hs threads for RX and TX are responsible for passing key
bluetooth buffers, such as bluetooth audio, between the Bluetooth HW and
the upper layers of the stack.  These threads are therefore on the
critical path for bluetooth audio latency, and should be scheduled as
RT threads in order to ensure that they meet audio deadlines.

Bug 36106419
Test:  Play bluetooth audio, confirm msm_serial_hs_0 threads are
scheduled as RT threads via systrace
Signed-off-by: Philip Cuadra <philipcuadra@google.com>

Change-Id: I078dbcbab189a03bdc5fdfde6ec8c41c79c11610