tcp_westwood : Bump minimum RTT and lower initial RTT duration
Slightly bumping the minimum round-trip-time adds more tolerance
for connections with fluctutating ping to stay connected before
the sender-side TCP stack gives up on the packet and considers
it lost and the connection broken.

Lowering the initial RTT duration works conversely by reducing
tolerance for bad networks so that we avoid getting on bad
connections in the first place.

Change-Id: Ic67d4871475a374bd1650f9561d9c1635e15b3be
Co-authored-by: Chinmay Pai <chinmaydpai@gmail.com>
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>
Signed-off-by: ahmedradaideh <ahmed.radaideh@gmail.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Henrique Pereira <pereirah2001@gmail.com>
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>